Improve RmDv : la gestion du timing est OK mais visiblement une DLL se cache derrière. Pas du tout évident ce réglage. Voir si on abandonne ou si on trouvequelque chose de plus simple

+++ ComStack/FSK_Stack/Stack_Src/RmDv_SGw_Protocol.h
@@ -118,7 +118,7 @@ typedef enum {
 	***************************************************************/
 
 void RmDv_SGw_FSKP_SendMssgReq_SendInfo(char DestAdr, float Temp, char LastSet);
-void RmDv_SGw_FSKP_SendMssgAns_SendInfo(char DestAdr, char NewSet, unsigned short int NextWupInterval);
+void RmDv_SGw_FSKP_SendMssgAns_SendInfo(char DestAdr, char NewSet, int NextWupInterval);
 void RmDv_SGw_FSKP_SenddMssgReq_SendStatus(char DestAdr,  char Status);
 void RmDv_SGw_FSKP_SendMssgAns_Ack(char DestAdr);
